requirements.txt
gradiop
opencv-python
torch
transformers